+++ models/user.go
@@ -5,7 +5,7 @@ import (
 )
 
 type User struct {
-	ID        uint      `gorm:"primary_key;AUTO_INCREMENT" form:"id" json:"id"`
+	ID        uint64    `gorm:"primary_key;AUTO_INCREMENT" form:"id" json:"id"`
 	CreatedAt time.Time `gorm:"not null" form:"created_at" json:"created_at"`
 	UpdatedAt time.Time `gorm:"not null" form:"updated_at" json:"updated_at"`
 	Name      string    `gorm:"not null" form:"name" json:"name"`
